lib: take the newer libc and libu sources from the ccc tree
These two libraries are maintained in ~/src/ccc and copied here, and
the copy had fallen behind by ten files.  This tree is the one that is
true, so the fixes come back to it; each of them was diagnosed over
there against a symptom seen here.

The stdio fixes are one bug with three faces.  fseek called _flsbuf
with one argument when it takes two, so the FILE * went in as the
character and the stream came off whatever was next on the stack; it
returned EOF and the lseek was never issued.  It is also the wrong
call - _flsbuf writes a whole BUFSIZ because the only caller that
wants it has a full buffer - and fflush, which writes the bytes that
are there, is what it should have been.  Then a seek leaves a
read-write stream with neither direction bit set, because the seek is
what makes the next operation free to be either, and both filbuf and
fgetc refused that instead of taking it as a read: every read after a
seek on a "w+" stream returned EOF, and fgetc latched _IOEOF on the
way out so the stream stayed dead afterwards.  asz is what found all
three - it assembles into a temp file, seeks back to the start and
copies it into the object, and it was writing objects that were a
bare sixteen byte header with none of the code in them.

The rest:

  fputc.s	drops the CP/M text mode \r ahead of every \n.  This
		is a Unix stdio and \n is a byte.
  freopen.c	takes the mode modifiers in either order, so that
		"w+b" and "wb+" both mean what they say.
  doprnt.c	an unsigned conversion went to unsigned long through
		a sign extension, so every value with the top bit set
		printed with ffff in front of it.  The linker's load
		map called sexit "FFFF9A41".
  link.s	restored the stack by four when three pops had moved
		it by six, so ret took the saved pointer for a return
		address: link returned into a filename.  What it
		looked like was the compiler driver restarting itself
		in the middle of a -O compile and announcing "no
		input files specified" with the assembler temporary
		as its program name.
  kill.s	the same off-by-one word, found beside it.  Nothing
		here calls kill yet, so it had never had the chance.

malloc.c changes a contract and not only an implementation.  malloc
does not come back empty any more: it writes "out of memory" and
exits, and __malloc beside it is the entry point for a caller that
genuinely can carry on without the memory.  The reason is that a null
return here is not a null dereference and a core file - there is no
core file.  The first write through the null lands on page zero, on
top of the rst 08 syscall trap, so the next system call does not trap
but runs forward through the zeroed page into 0x0100, which is the
entry point, and the program starts over.  cmd/s and cmd/vi have three
call sites between them that test malloc for null; those tests are
dead now rather than wrong, and vi's "not enough memory" becomes
malloc's message instead.

swdispw.s is new here.  It is the code generator's dispatch helper for
a switch whose case values do not fit in a byte - case S_IFDIR and its
neighbours - and this tree carried swdisp.s, the byte one, without it.
Nothing has needed it yet, but whether it is needed is the compiler's
decision and not the library's, and it is a separate object precisely
so that a program without such a switch does not link it.

rename.c is ccc's rewrite.  It lives in libu there and in libc here,
and that question is left open: the argument for libu is that micronix
may grow a rename system call, and when it does this becomes an .s
stub shaped like link.s with nothing above it changing.

_signal.s is deliberately NOT taken.  This tree is the one that is
ahead of ccc there: the .global __signal and .global _jtab
declarations were added here, and without them every symbol in the
file is local, the object defines nothing and signal() cannot be
linked at all - which is what cmd/sh/NOTES describes.  ccc's copy is
still the December one.  That fix wants to go the other way.

The cross build is clean and all thirteen commands link.  libu's own
check target still finds __signal and _jtab exported.

+++ src/micronix/lib/libc/doprnt.c
@@ -172,7 +172,15 @@ dostring:
 				width = _fnum((char *)a, prec, width, sign, pputc);
 				a += FLTSIZE/sizeof(*a);
 			} else {
-				width = _pnum((len == sizeof(int)/sizeof *a ? (sign ? (long)*a : (unsigned long)*a) : *(long *)a), prec, width, sign, base, pputc);
+				/*
+				 * (unsigned)*a first: a is int *, so going
+				 * straight to unsigned long sign-extends,
+				 * and every unsigned conversion of a value
+				 * with the top bit set printed eight digits
+				 * with ffff in front.  The linker's load map
+				 * called sexit "FFFF9A41".
+				 */
+				width = _pnum((len == sizeof(int)/sizeof *a ? (sign ? (long)*a : (unsigned long)(unsigned int)*a) : *(long *)a), prec, width, sign, base, pputc);
 				a += len;
 			}
 			while(left-- > width)
